# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Calculadora de Débitos Judiciais",
    "version": "2.3.2"
  },
  "paths": {
...
    "/api/v2/qualidade/revisoes": {
      "post": {
        "tags": [
          "Qualidade IA"
        ],
        "summary": "Capture Review",
        "description": "Registra IA × humano sem sobrescrever a predição original.",
        "operationId": "capture_review_api_v2_qualidade_revisoes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSnapshotInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCaptureResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/qualidade/feedback": {
      "get": {
        "tags": [
          "Qualidade IA"
        ],
        "summary": "Feedback Page",
        "operationId": "feedback_page_api_v2_qualidade_feedback_get",
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Pagina"
            }
          },
          {
            "name": "tamanho",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Tamanho"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "pending",
              "title": "Status"
            }
          },
          {
            "name": "campo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Campo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/qualidade/feedback/{feedback_id}/curadoria": {
      "post": {
        "tags": [
          "Qualidade IA"
        ],
        "summary": "Curate Feedback",
        "operationId": "curate_feedback_api_v2_qualidade_feedback__feedback_id__curadoria_post",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feedback Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCurationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackEventRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/qualidade/finops": {
      "get": {
        "tags": [
          "Qualidade IA"
        ],
        "summary": "Finops",
        "operationId": "finops_api_v2_qualidade_finops_get",
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 30,
              "title": "Dias"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinOpsSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FeedbackCurationInput": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Status"
          },
          "motivo_codigo": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "confirmado_sem_alteracao",
                  "decisao_posterior_prevalece",
                  "documento_incorreto",
                  "pagina_incorreta",
                  "valor_interpretado_incorretamente",
                  "data_interpretada_incorretamente",
                  "regra_nao_se_aplica",
                  "campo_ausente_no_documento",
                  "contrato_incorreto",
                  "parcela_associada_ao_contrato_errado",
                  "classificacao_documento_incorreta",
                  "ambiguidade_documental",
                  "adicao_manual",
                  "nao_informado",
                  "outro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Codigo"
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FeedbackCurationInput"
      },
      "FeedbackEventRecord": {
        "properties": {
          "identificador": {
            "type": "string",
            "title": "Identificador"
          },
          "numero_processo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Processo"
          },
          "rascunho_id": {
            "type": "string",
            "title": "Rascunho Id"
          },
          "extracao_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracao Id"
          },
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "acao": {
            "type": "string",
            "enum": [
              "confirmed",
              "corrected",
              "removed",
              "added",
              "not_found",
              "ambiguous"
            ],
            "title": "Acao"
          },
          "motivo_codigo": {
            "type": "string",
            "enum": [
              "confirmado_sem_alteracao",
              "decisao_posterior_prevalece",
              "documento_incorreto",
              "pagina_incorreta",
              "valor_interpretado_incorretamente",
              "data_interpretada_incorretamente",
              "regra_nao_se_aplica",
              "campo_ausente_no_documento",
              "contrato_incorreto",
              "parcela_associada_ao_contrato_errado",
              "classificacao_documento_incorreta",
              "ambiguidade_documental",
              "adicao_manual",
              "nao_informado",
              "outro"
            ],
            "title": "Motivo Codigo"
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario"
          },
          "valor_modelo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Modelo"
          },
          "valor_humano": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Humano"
          },
          "documento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documento"
          },
          "pagina": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagina"
          },
          "evidencia_modelo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidencia Modelo"
          },
          "tipo_documento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo Documento"
          },
          "versao_prompts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versao Prompts"
          },
          "modelo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelo"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version"
          },
          "revisor_pseudonimo": {
            "type": "string",
            "title": "Revisor Pseudonimo"
          },
          "criado_em": {
            "type": "string",
            "title": "Criado Em"
          },
          "status_curadoria": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "title": "Status Curadoria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identificador",
          "rascunho_id",
          "campo",
          "acao",
          "motivo_codigo",
          "pipeline_version",
          "revisor_pseudonimo",
          "criado_em",
          "status_curadoria"
        ],
        "title": "FeedbackEventRecord"
      },
...
      "FeedbackPage": {
        "properties": {
          "itens": {
            "items": {
              "$ref": "#/components/schemas/FeedbackEventRecord"
            },
            "type": "array",
            "title": "Itens"
          },
          "pagina": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Pagina"
          },
          "tamanho_pagina": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Tamanho Pagina"
          },
          "total_itens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Itens"
          },
          "total_paginas": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Paginas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "itens",
          "pagina",
          "tamanho_pagina",
          "total_itens",
          "total_paginas"
        ],
        "title": "FeedbackPage"
      },
...
      "FinOpsStageMetric": {
        "properties": {
          "etapa": {
            "type": "string",
            "title": "Etapa"
          },
          "chamadas_api": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Chamadas Api"
          },
          "falhas_api": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Falhas Api",
            "default": 0
          },
          "acertos_cache": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Acertos Cache"
          },
          "caracteres_entrada": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Caracteres Entrada"
          },
          "caracteres_saida": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Caracteres Saida"
          },
          "tokens_reais": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Reais"
          },
          "tokens_estimados": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tokens Estimados"
          },
          "custo_estimado_usd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custo Estimado Usd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "etapa",
          "chamadas_api",
          "acertos_cache",
          "caracteres_entrada",
          "caracteres_saida",
          "tokens_estimados"
        ],
        "title": "FinOpsStageMetric"
      },
      "FinOpsSummary": {
        "properties": {
          "periodo_dias": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Periodo Dias"
          },
          "chamadas_api": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Chamadas Api"
          },
          "falhas_api": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Falhas Api",
            "default": 0
          },
          "acertos_cache": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Acertos Cache"
          },
          "taxa_cache": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
            "title": "Taxa Cache"
          },
          "caracteres_entrada": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Caracteres Entrada"
          },
          "caracteres_saida": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Caracteres Saida"
          },
          "tokens_reais": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Reais"
          },
          "tokens_estimados": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tokens Estimados"
          },
          "custo_estimado_usd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custo Estimado Usd"
          },
          "origem_tokens": {
            "type": "string",
            "enum": [
              "gateway",
              "estimativa_local",
              "mista",
              "cache",
              "indisponivel"
            ],
            "title": "Origem Tokens"
          },
          "economia_chamadas_cache": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Economia Chamadas Cache"
          },
          "orcamento_mensal_usd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Mensal Usd"
          },
          "consumo_mes_estimado_usd": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumo Mes Estimado Usd"
          },
          "percentual_orcamento": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentual Orcamento"
          },
          "status_orcamento": {
            "type": "string",
            "enum": [
              "not_configured",
              "unavailable",
              "ok",
              "warning",
              "exceeded"
            ],
            "title": "Status Orcamento",
            "default": "not_configured"
          },
          "por_etapa": {
            "items": {
              "$ref": "#/components/schemas/FinOpsStageMetric"
            },
            "type": "array",
            "title": "Por Etapa"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "periodo_dias",
          "chamadas_api",
          "acertos_cache",
          "taxa_cache",
          "caracteres_entrada",
          "caracteres_saida",
          "tokens_estimados",
          "origem_tokens",
          "economia_chamadas_cache"
        ],
        "title": "FinOpsSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Installment-Input": {
        "properties": {
          "data": {
            "type": "string",
            "format": "date",
            "title": "Data"
          },
          "valor_singelo": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Valor Singelo"
          },
          "descricao": {
            "type": "string",
            "maxLength": 500,
            "title": "Descricao",
            "default": ""
          },
          "verba_tipo": {
            "type": "string",
            "enum": [
              "dano_material",
              "dano_moral",
              "honorarios",
              "custas"
            ],
            "title": "Verba Tipo"
          },
          "multiplicador": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  1,
                  2
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Multiplicador"
          },
          "origem": {
            "type": "string",
            "enum": [
              "informada",
              "honorarios_dano_moral"
            ],
            "title": "Origem",
            "default": "informada"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "valor_singelo",
          "verba_tipo"
        ],
        "title": "Installment",
        "description": "Parcela revisada e congelada dentro de uma versão."
      },
...
      "ReviewCaptureResult": {
        "properties": {
          "eventos_criados": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Eventos Criados"
          },
          "confirmados": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Confirmados"
          },
          "corrigidos": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Corrigidos"
          },
          "pendentes_curadoria": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pendentes Curadoria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "eventos_criados",
          "confirmados",
          "corrigidos",
          "pendentes_curadoria"
        ],
        "title": "ReviewCaptureResult"
      },
      "ReviewSnapshotInput": {
        "properties": {
          "origem_calculo": {
            "type": "string",
            "enum": [
              "manual",
              "processo"
            ],
            "title": "Origem Calculo"
          },
          "numero_processo": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9][0-9.\\-]{0,39}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Processo"
          },
          "rascunho_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 8,
            "title": "Rascunho Id"
          },
          "extracao_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracao Id"
          },
          "parametros": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Parametros"
          },
          "parametros_por_dano": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "type": "object",
            "title": "Parametros Por Dano"
          },
          "parcelas": {
            "items": {
              "$ref": "#/components/schemas/Installment-Input"
            },
            "type": "array",
            "title": "Parcelas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origem_calculo",
          "rascunho_id"
        ],
        "title": "ReviewSnapshotInput",
        "description": "Snapshot do estado revisado usado para comparar IA × humano no servidor."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}